pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
alembic = "^1.14.1"
jinja2 = "^3.1.5"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
